v2
Changed the way how the item names are read from game data and how the search bar code works. 
Now sub-items like colored wool, stairs, and other items and blocks based on the damage index (meta) can be searched by the name of the sub-item directly like color, material, etc. instead of searching for the name of the root item. 
This also fixes displayed names for blocks and items if in the code their class is based on extending another block's or item's class.
Fixed how background rectangle behind item names is rendered in inventory.
Refactoring.